Move the aliases and base funcionality from the NextcloudStorage to Storage
