+++ modules/storages/spec/models/storages/nextcloud_storage_spec.rb
@@ -181,28 +181,6 @@ RSpec.describe Storages::NextcloudStorage do
     end
   end
 
-  shared_examples 'a stored boolean attribute' do |attribute|
-    it "#{attribute} has a default value of false" do
-      expect(storage.public_send(:"#{attribute}?")).to be(false)
-    end
-
-    ['1', 'true', true].each do |boolean_like|
-      context "with truthy value #{boolean_like}" do
-        it "sets #{attribute} to true" do
-          storage.public_send(:"#{attribute}=", boolean_like)
-          expect(storage.public_send(attribute)).to be(true)
-        end
-      end
-    end
-
-    it "#{attribute} can be set to true" do
-      storage.public_send(:"#{attribute}=", true)
-
-      expect(storage.public_send(attribute)).to be(true)
-      expect(storage.public_send(:"#{attribute}?")).to be(true)
-    end
-  end
-
   describe '#username' do
     it_behaves_like 'a stored attribute with default value', :username, 'OpenProject'
   end
@@ -213,30 +191,6 @@ RSpec.describe Storages::NextcloudStorage do
 
   describe '#group_folder' do
     it_behaves_like 'a stored attribute with default value', :group_folder, 'OpenProject'
-  end
-
-  describe '#automatically_managed' do
-    it_behaves_like 'a stored boolean attribute', :automatically_managed
-  end
-
-  describe '#automatic_management_enabled?' do
-    context 'when automatic management enabled is true' do
-      let(:storage) { build(:nextcloud_storage, automatic_management_enabled: true) }
-
-      it { expect(storage).to be_automatic_management_enabled }
-    end
-
-    context 'when automatic management enabled is false' do
-      let(:storage) { build(:nextcloud_storage, automatic_management_enabled: false) }
-
-      it { expect(storage).not_to be_automatic_management_enabled }
-    end
-
-    context 'when automatic management enabled is nil' do
-      let(:storage) { build(:nextcloud_storage, automatic_management_enabled: nil) }
-
-      it { expect(storage.automatic_management_enabled?).to be(false) }
-    end
   end
 
   describe '#automatic_management_new_record?' do
